fix(cli): apply toRawUrl on --import url fetches
Share the GUI blob→raw rewrite so GitHub/GitLab/Bitbucket URLs
fetch Markdown instead of HTML.

Closes #197
